Set up rough player scripting
- Add start and update script function
- Add UUID entity map
- Change UUID to uint32 to work better with Lua
- All Scripts updated to new UUID

+++ Rhombus/src/Rhombus/Core/UUID.cpp
@@ -9,14 +9,14 @@ namespace rhombus
 {
 	static std::random_device s_RandomDevice;
 	static std::mt19937_64 s_Engine(s_RandomDevice());
-	static std::uniform_int_distribution<uint64_t> s_UniformDistribution;
+	static std::uniform_int_distribution<uint32_t> s_UniformDistribution;
 
 	UUID::UUID()
 		: m_UUID(s_UniformDistribution(s_Engine))
 	{
 	}
 
-	UUID::UUID(uint64_t uuid)
+	UUID::UUID(uint32_t uuid)
 		: m_UUID(uuid)
 	{
 	}
