+++ Rhombus/src/Rhombus/Core/UUID.h
@@ -6,13 +6,13 @@ namespace rhombus
 	{
 	public:
 		UUID();
-		UUID(uint64_t uuid);
+		UUID(uint32_t uuid);
 		UUID(const UUID&) = default;
 
-		operator uint64_t() const{ return m_UUID; }
+		operator uint32_t() const{ return m_UUID; }
 
 	private:
-		uint64_t m_UUID;
+		uint32_t m_UUID;
 	};
 }
 
@@ -25,7 +25,7 @@ namespace std
 	{
 		std::size_t operator()(const rhombus::UUID& uuid) const
 		{
-			return (uint64_t)uuid;
+			return (uint32_t)uuid;
 		}
 	};
 }
